[blame: fix miss-aligned file and blame]
issue:

Blame buffer information does not belong to lines in the file.

reason:

The blame plugin does not account for folded regions. Hence, if a file
contains such regions, all lines following the first fold are not in
line with the blame buffer.

fix:

Open all folded regions in a file before populating the blame buffer.